engines/sg: extract BE accessor helpers into shared sg.h
The six static inline byte-order helpers (sgio_get/set_be16/32/64)
were defined privately inside sg.c, making them inaccessible to other
engines that also need to construct or parse SCSI CDBs.

Move them into a new engines/sg.h so that any engine speaking SCSI can
include the header directly rather than reimplementing the same
byte-swap wrappers independently.  sg.c gains an include of sg.h in
place of the now-removed definitions; no behaviour is changed.
