+++ engines/sg.c
@@ -57,6 +57,7 @@
 
 #include "../fio.h"
 #include "../optgroup.h"
+#include "sg.h"
 
 #ifdef FIO_HAVE_SGIO
 
@@ -216,42 +217,6 @@ struct sgio_data {
 	unsigned int *trim_queue_map;
 #endif
 };
-
-static inline uint16_t sgio_get_be16(uint8_t *buf)
-{
-	return be16_to_cpu(*((uint16_t *) buf));
-}
-
-static inline uint32_t sgio_get_be32(uint8_t *buf)
-{
-	return be32_to_cpu(*((uint32_t *) buf));
-}
-
-static inline uint64_t sgio_get_be64(uint8_t *buf)
-{
-	return be64_to_cpu(*((uint64_t *) buf));
-}
-
-static inline void sgio_set_be16(uint16_t val, uint8_t *buf)
-{
-	uint16_t t = cpu_to_be16(val);
-
-	memcpy(buf, &t, sizeof(uint16_t));
-}
-
-static inline void sgio_set_be32(uint32_t val, uint8_t *buf)
-{
-	uint32_t t = cpu_to_be32(val);
-
-	memcpy(buf, &t, sizeof(uint32_t));
-}
-
-static inline void sgio_set_be64(uint64_t val, uint8_t *buf)
-{
-	uint64_t t = cpu_to_be64(val);
-
-	memcpy(buf, &t, sizeof(uint64_t));
-}
 
 static inline bool sgio_unbuffered(struct thread_data *td)
 {
